+++ packages/wallet-connectors/src/WalletConnect.ts
@@ -8,13 +8,20 @@ import {
     HttpProvider,
     InitContractPayload,
     JsonRpcClient,
-    SchemaVersion,
     serializeInitContractParameters,
+    serializeTypeValue,
     serializeUpdateContractParameters,
     toBuffer,
     UpdateContractPayload,
 } from '@concordium/web-sdk';
-import { Network, WalletConnection, WalletConnectionDelegate, WalletConnector } from './WalletConnection';
+import {
+    Network,
+    TypedSmartContractParameters,
+    WalletConnection,
+    WalletConnectionDelegate,
+    WalletConnector,
+} from './WalletConnection';
+import { UnreachableCaseError } from './error';
 
 const WALLET_CONNECT_SESSION_NAMESPACE = 'ccd';
 
@@ -71,77 +78,83 @@ function accountTransactionPayloadToJson(data: AccountTransactionPayload) {
     });
 }
 
+function serializeInitContractParam(initName: string, typedParams: TypedSmartContractParameters | undefined) {
+    if (!typedParams) {
+        return toBuffer('');
+    }
+    const { parameters, schema } = typedParams;
+    switch (schema.type) {
+        case 'ModuleSchema':
+            return serializeInitContractParameters(initName, parameters, schema.value, schema.version);
+        case 'ParameterSchema':
+            return serializeTypeValue(parameters, schema.value);
+        default:
+            throw new UnreachableCaseError('schema', schema);
+    }
+}
+
+function serializeUpdateContractMessage(
+    contractName: string,
+    entrypointName: string,
+    typedParams: TypedSmartContractParameters | undefined
+) {
+    if (!typedParams) {
+        return toBuffer('');
+    }
+    const { parameters, schema } = typedParams;
+    switch (schema.type) {
+        case 'ModuleSchema':
+            return serializeUpdateContractParameters(
+                contractName,
+                entrypointName,
+                parameters,
+                schema.value,
+                schema.version
+            );
+        case 'ParameterSchema':
+            return serializeTypeValue(parameters, schema.value);
+        default:
+            throw new UnreachableCaseError('schema', schema);
+    }
+}
+
 /**
- * Encode parameters into appropriate payload field ('payload.param' for 'InitContract' and 'payload.message' for 'Update').
- * The 'parameters' and 'schema' parameters must be not undefined for these transaction types.
- * The payload field must be not already set as that would indicate that the caller thought that was the right way to pass them.
+ * Serialize parameters into appropriate payload field ('payload.param' for 'InitContract' and 'payload.message' for 'Update').
+ * This payload field must be not already set as that would indicate that the caller thought that was the right way to pass them.
  * @param type Type identifier of the transaction.
  * @param payload Payload of the transaction. Must not include the fields 'param' and 'message' for transaction types 'InitContract' and 'Update', respectively.
- * @param parameters Contract invocation parameters. Must be provided for 'InitContract' or 'Update' transactions and omitted otherwise.
- * @param schema Schema for the contract invocation parameters. Must be provided for 'InitContract' or 'Update' transactions and omitted otherwise.
- * @param schemaVersion Version of the provided schema.
+ * @param typedParams Contract invocation parameters and associated schema. May be provided optionally provided for transactions of type 'InitContract' or 'Update'.
  */
-function encodePayloadParameters(
+function serializePayloadParameters(
     type: AccountTransactionType,
     payload: AccountTransactionPayload,
-    parameters?: Record<string, unknown>,
-    schema?: string,
-    schemaVersion?: SchemaVersion
-) {
+    typedParams: TypedSmartContractParameters | undefined
+): AccountTransactionPayload {
     switch (type) {
         case AccountTransactionType.InitContract: {
-            if (parameters === undefined) {
-                throw new Error(`parameters provided for 'InitContract' transaction must be not undefined`);
-            }
-            if (schema === undefined) {
-                throw new Error(`schema provided for 'InitContract' transaction must be not undefined`);
-            }
             const initContractPayload = payload as InitContractPayload;
             if (initContractPayload.param) {
                 throw new Error(`'param' field of 'InitContract' parameters must be empty`);
             }
             return {
                 ...payload,
-                param: serializeInitContractParameters(
-                    initContractPayload.initName,
-                    parameters,
-                    toBuffer(schema, 'base64'),
-                    schemaVersion
-                ),
-            } as InitContractPayload;
+                param: serializeInitContractParam(initContractPayload.initName, typedParams),
+            };
         }
         case AccountTransactionType.Update: {
-            if (parameters === undefined) {
-                throw new Error(`parameters provided for 'Update' transaction must be not undefined`);
-            }
-            if (schema === undefined) {
-                throw new Error(`schema provided for 'Update' transaction must be not undefined`);
-            }
             const updateContractPayload = payload as UpdateContractPayload;
             if (updateContractPayload.message) {
                 throw new Error(`'message' field of 'Update' parameters must be empty`);
             }
-            const [contractName, receiveName] = updateContractPayload.receiveName.split('.');
+            const [contractName, entrypointName] = updateContractPayload.receiveName.split('.');
             return {
                 ...payload,
-                message: serializeUpdateContractParameters(
-                    contractName,
-                    receiveName,
-                    parameters,
-                    toBuffer(schema, 'base64'),
-                    schemaVersion
-                ),
-            } as UpdateContractPayload;
+                message: serializeUpdateContractMessage(contractName, entrypointName, typedParams),
+            };
         }
         default: {
-            if (parameters !== undefined) {
-                throw new Error(`parameters provided for '${type}' transaction must be undefined`);
-            }
-            if (schema !== undefined) {
-                throw new Error(`schema provided for '${type}' transaction must be undefined`);
-            }
-            if (schemaVersion !== undefined) {
-                throw new Error(`schema version provided for '${type}' transaction must be undefined`);
+            if (typedParams) {
+                throw new Error(`'typedParams' must not be provided for transaction of type '${type}'`);
             }
             return payload;
         }
@@ -204,17 +217,13 @@ export class WalletConnectConnection implements WalletConnection {
         accountAddress: string,
         type: AccountTransactionType,
         payload: AccountTransactionPayload,
-        parameters?: Record<string, unknown>,
-        schema?: string,
-        schemaVersion?: SchemaVersion
+        typedParams?: TypedSmartContractParameters
     ) {
         const params = {
             type: AccountTransactionType[type],
             sender: accountAddress,
-            payload: accountTransactionPayloadToJson(
-                encodePayloadParameters(type, payload, parameters, schema, schemaVersion)
-            ),
-            schema,
+            payload: accountTransactionPayloadToJson(serializePayloadParameters(type, payload, typedParams)),
+            schema: typedParams?.schema,
         };
         try {
             const { hash } = (await this.connector.client.request({